Split litter_pools submodule into seperate carbon and chemistry submodules
