+++ docs/source/_toc.yaml
@@ -171,6 +171,10 @@ subtrees:
           - file: api/models/litter
             title: The litter model
             entries:
+              - file: api/models/litter/carbon
+                title: The litter carbon submodule
+              - file: api/models/litter/chemistry
+                title: The litter chemistry submodule
               - file: api/models/litter/constants
                 title: The constants submodule
               - file: api/models/litter/env_factors
@@ -179,8 +183,6 @@ subtrees:
                 title: The input_partition submodule
               - file: api/models/litter/litter_model
                 title: The litter_model submodule
-              - file: api/models/litter/litter_pools
-                title: The litter_pools submodule
           - file: api/models/soil
             title: The soil model
             entries:
