+++ docs/source/api/models/litter/carbon.md
@@ -1,0 +1,23 @@
+---
+jupytext:
+  cell_metadata_filter: -all
+  formats: md:myst
+  main_language: python
+  text_representation:
+    extension: .md
+    format_name: myst
+    format_version: 0.13
+    jupytext_version: 1.13.8
+kernelspec:
+  display_name: Python 3 (ipykernel)
+  language: python
+  name: python3
+---
+
+# API documentation for the {mod}`~virtual_ecosystem.models.litter.carbon` module
+
+```{eval-rst}
+.. automodule:: virtual_ecosystem.models.litter.carbon
+    :autosummary:
+    :members:
+```
